* 'master' of github.com:tensorflow/tensorflow: (115 commits)
  Add base_layer_util to frozen_keras.
  Disable stacktrace tests on windows.
  Add cross host send/recv to PyLocalClient. Not implemented for now.
  Disable failing placer_test on windows.
  Add CreateTokenOp to HLO dialect.
  Adds support for depth_multiplier case when DepthwiseConv resolves to Conv
  Make test pass in 2.0
  Internal visibility change.
  Changed accumulator precision for devices that does not support round to nearest.
  Add API to overwrite the float interface of quantized tflite model. Currently supports int8 and uint8.
  Add more tests for LSTM op for quantized case.
  [tf.slice()] Optimize the shared validation code.
  [MLIR][XLA] Lower xla_hlo.rsqrt from HLO -> LHLO -> Linalg.
  [XLA][MLIR] Add ReduceWindowOp to LHLO dialect.
  Change passing in LaunchOp where only body of LaunchOp is needed
  Bump LLVM version.
  Updated kkb@ TODOs (ldap changed)
  Go: Update generated wrapper functions for TensorFlow ops.
  Go: Update generated wrapper functions for TensorFlow ops.
  Add cross build parameters to android build guide
  ...

+++ tensorflow/core/kernels/data/experimental/rebatch_dataset_op.cc
@@ -62,7 +62,8 @@ class RebatchDatasetOp : public UnaryDatasetOpKernel {
           output_types_(output_types),
           output_shapes_(output_shapes),
           traceme_metadata_(
-              {{"num_replicas", strings::Printf("%lld", num_replicas)}}) {
+              {{"num_replicas", strings::Printf("%lld", static_cast<long long>(
+                                                            num_replicas))}}) {
       input_->Ref();
     }
